brand_monitoring/requirements.txt
newspaper3k
googlesearch-python
feedparser
pandas
textblob
lxml[html_clean]
